refactor(frontend): extract IngestionWizard step components (#256)
Extract PasteStep, UploadingStep, PreviewStep, EditingStep, and
ConfirmingStep into frontend/src/components/ingestion/. Keep state,
polling, persistence, and raw fetch API helpers in IngestionWizard.tsx.
Move mapPreviewToFormData to a shared helper so PreviewStep can reuse it
without a circular dependency.
