+++ frontend/src/components/IngestionWizard.tsx
@@ -1,11 +1,13 @@
 import { useCallback, useEffect, useRef, useState } from "react";
 import type { ChangeEvent, ClipboardEvent } from "react";
-import { GraphSandbox } from "./GraphSandbox";
-import { TagInput } from "./TagInput";
-import { LatexText } from "./LatexText";
-import { parseOptions } from "./AnswerInput";
 import { api } from "@/api/client";
 import { useTagSuggestions } from "@/hooks/useTagSuggestions";
+import { PasteStep } from "./ingestion/PasteStep";
+import { UploadingStep } from "./ingestion/UploadingStep";
+import { PreviewStep } from "./ingestion/PreviewStep";
+import { EditingStep } from "./ingestion/EditingStep";
+import { ConfirmingStep } from "./ingestion/ConfirmingStep";
+import { mapPreviewToFormData } from "./ingestion/mapPreviewToFormData";
 
 interface PreviewSourceImage {
   bucket: string;
@@ -96,17 +98,6 @@ function normalizePreviewResponse(data: unknown): IngestionPreview {
     createdAt: candidate.createdAt ?? "",
     updatedAt: candidate.updatedAt ?? "",
     expiresAt: candidate.expiresAt ?? "",
-  };
-}
-
-function mapPreviewToFormData(preview: IngestionPreview) {
-  return {
-    text: preview.draft.text || "",
-    problemType: preview.draft.problemType || "",
-    graphDsl: preview.draft.graphDsl || "",
-    correctAnswer: preview.draft.correctAnswer || "",
-    tags: preview.draft.tags,
-    subject: preview.draft.subject || "math",
   };
 }
 
@@ -224,941 +215,6 @@ function useDebouncedCallback<T extends (...args: unknown[]) => unknown>(
       }, delay);
     },
     [callback, delay]
-  );
-}
-
-// ── Step sub-components ──────────────────────────────────────────────
-
-interface PasteStepProps {
-  onPaste: (event: ClipboardEvent) => void;
-  fileInputRef: React.RefObject<HTMLInputElement>;
-  onFileSelection: (event: ChangeEvent<HTMLInputElement>) => void;
-  error: string;
-}
-
-function PasteStep({ onPaste, fileInputRef, onFileSelection, error }: PasteStepProps) {
-  return (
-    <div
-      style={{
-        padding: "48px 32px",
-        textAlign: "center",
-        border: "2px dashed var(--color-border-muted)",
-        borderRadius: "8px",
-        backgroundColor: "var(--color-surface-muted)",
-      }}
-      onPaste={onPaste}
-      tabIndex={0}
-      role="region"
-      aria-label="Paste image area"
-    >
-      <div style={{ marginBottom: "16px", fontSize: "48px" }}>📋</div>
-      <h3 style={{ margin: "0 0 8px", fontSize: "18px", fontWeight: 600 }}>
-        Paste an Image
-      </h3>
-      <p style={{ margin: 0, color: "var(--color-text-muted)", fontSize: "14px" }}>
-        Copy an image and paste it here (Ctrl+V or Cmd+V)
-      </p>
-      <div style={{ marginTop: "16px" }}>
-        <button
-          type="button"
-          onClick={() => fileInputRef.current?.click()}
-          style={{
-            padding: "8px 16px",
-            backgroundColor: "var(--color-surface)",
-            color: "var(--color-text)",
-            border: "1px solid var(--color-border-muted)",
-            borderRadius: "4px",
-            cursor: "pointer",
-          }}
-        >
-          Choose Image File
-        </button>
-        <input
-          ref={fileInputRef}
-          type="file"
-          accept="image/*"
-          onChange={(e) => void onFileSelection(e)}
-          style={{ display: "none" }}
-        />
-      </div>
-      {error && (
-        <div
-          style={{
-            marginTop: "16px",
-            padding: "12px 16px",
-            backgroundColor: "var(--color-danger-bg)",
-            borderRadius: "6px",
-            color: "var(--color-text-danger)",
-            fontSize: "14px",
-          }}
-        >
-          {error}
-        </div>
-      )}
-    </div>
-  );
-}
-
-interface UploadingStepProps {
-  uploadProgress: number;
-}
-
-function UploadingStep({ uploadProgress }: UploadingStepProps) {
-  return (
-    <div style={{ padding: "48px 32px", textAlign: "center" }}>
-      <div style={{ marginBottom: "16px", fontSize: "48px" }}>⏳</div>
-      <h3 style={{ margin: "0 0 8px", fontSize: "18px", fontWeight: 600 }}>
-        Uploading...
-      </h3>
-      <div
-        style={{
-          width: "100%",
-          height: "8px",
-          backgroundColor: "var(--color-border)",
-          borderRadius: "4px",
-          overflow: "hidden",
-          marginTop: "16px",
-        }}
-      >
-        <div
-          style={{
-            width: `${uploadProgress}%`,
-            height: "100%",
-            backgroundColor: "var(--color-primary)",
-            transition: "width 0.3s ease",
-          }}
-        />
-      </div>
-      <p style={{ margin: "8px 0 0", color: "var(--color-text-muted)", fontSize: "14px" }}>
-        {uploadProgress}%
-      </p>
-    </div>
-  );
-}
-
-interface PreviewStepProps {
-  preview: IngestionPreview | null;
-  isLoading: boolean;
-  onRetry: () => void;
-  setFormData: React.Dispatch<React.SetStateAction<WizardFormData>>;
-  setCurrentStep: (step: WizardStep) => void;
-  error: string;
-  helperFailureSubject: string;
-  onHelperFailureSubjectChange: (subject: string) => void;
-}
-
-function PreviewStep({ preview, isLoading, onRetry, setFormData, setCurrentStep, error, helperFailureSubject, onHelperFailureSubjectChange }: PreviewStepProps) {
-  return (
-    <div style={{ padding: "32px" }}>
-      <h3 style={{ margin: "0 0 16px", fontSize: "18px", fontWeight: 600 }}>
-        Processing Image
-      </h3>
-      {preview?.status === "extracting" && (
-        <div style={{ textAlign: "center", padding: "32px" }}>
-          <div style={{ fontSize: "32px", marginBottom: "16px" }}>🤖</div>
-          <p style={{ color: "var(--color-text-muted)" }}>
-            AI is analyzing the image and extracting problem data...
-          </p>
-          <div
-            style={{
-              width: "40px",
-              height: "40px",
-              border: "3px solid var(--color-border)",
-              borderTopColor: "var(--color-primary)",
-              borderRadius: "50%",
-              margin: "16px auto",
-              animation: "spin 1s linear infinite",
-            }}
-          />
-        </div>
-      )}
-      {preview?.status === "vlm-failed" && (() => {
-        const helperDetection = preview?.helperDetection;
-        const isHelperFailure = helperDetection?.failureCode;
-        const failureCode = isHelperFailure
-          ? helperDetection.failureCode
-          : preview?.extraction?.failureCode;
-        const failureMessage = isHelperFailure
-          ? helperDetection.failureMessage
-          : preview?.extraction?.failureMessage;
-        return (
-        <div
-          style={{
-            padding: "16px",
-            backgroundColor: "var(--color-danger-bg)",
-            borderRadius: "6px",
-            marginBottom: "16px",
-          }}
-        >
-          <div style={{ color: "var(--color-text-danger)", fontWeight: 600, marginBottom: "8px" }}>
-            ⚠️ {isHelperFailure ? "Subject Classification Failed" : "Extraction Failed"}
-          </div>
-          <p style={{ color: "var(--color-text-danger-secondary)", fontSize: "14px", margin: "0 0 12px" }}>
-            {isHelperFailure
-              ? "The AI could not determine the subject (math or English). Please select the subject manually and retry."
-              : "The AI was unable to extract problem data from the image."}
-          </p>
-          {(failureCode || failureMessage) && (
-            <details style={{ marginBottom: "12px" }}>
-              <summary style={{ cursor: "pointer", color: "var(--color-text-danger-secondary)", fontSize: "13px" }}>
-                View error details
-              </summary>
-              <div
-                style={{
-                  marginTop: "8px",
-                  padding: "12px",
-                  backgroundColor: "var(--color-danger-border)",
-                  borderRadius: "4px",
-                  fontSize: "12px",
-                  fontFamily: "monospace",
-                }}
-              >
-                {failureCode && (
-                  <div style={{ marginBottom: "8px" }}>
-                    <span style={{ fontWeight: 600 }}>Code:</span> {failureCode}
-                  </div>
-                )}
-                {failureMessage && (
-                  <div>
-                    <span style={{ fontWeight: 600 }}>Message:</span>{" "}
-                    <pre style={{ margin: 0, whiteSpace: "pre-wrap", wordBreak: "break-word" }}>
-                      {failureMessage}
-                    </pre>
-                  </div>
-                )}
-              </div>
-            </details>
-          )}
-          {isHelperFailure && (
-            <div style={{ marginBottom: "12px" }}>
-              <label
-                htmlFor="helper-failure-subject"
-                style={{ display: "block", marginBottom: "4px", fontSize: "14px", color: "var(--color-text)" }}
-              >
-                Subject
-              </label>
-              <select
-                id="helper-failure-subject"
-                data-testid="helper-failure-subject-select"
-                value={helperFailureSubject}
-                onChange={(e) => onHelperFailureSubjectChange(e.target.value)}
-                disabled={isLoading}
-                style={{
-                  padding: "8px 12px",
-                  border: "1px solid var(--color-border)",
-                  borderRadius: "4px",
-                  backgroundColor: "var(--color-surface)",
-                  color: "var(--color-text)",
-                  fontSize: "14px",
-                }}
-              >
-                <option value="math">Math</option>
-                <option value="english">English</option>
-              </select>
-            </div>
-          )}
-          <div style={{ display: "flex", gap: "12px", flexWrap: "wrap" }}>
-            <button
-              onClick={onRetry}
-              disabled={isLoading}
-              style={{
-                padding: "8px 16px",
-                backgroundColor: "var(--color-danger)",
-                color: "white",
-                border: "none",
-                borderRadius: "4px",
-                cursor: isLoading ? "not-allowed" : "pointer",
-                opacity: isLoading ? 0.7 : 1,
-              }}
-            >
-              {isLoading ? "Retrying..." : "Try Again"}
-            </button>
-            <button
-              type="button"
-              onClick={() => {
-                if (preview) {
-                  setFormData(mapPreviewToFormData(preview));
-                  setCurrentStep("editing");
-                }
-              }}
-              style={{
-                padding: "8px 16px",
-                backgroundColor: "var(--color-surface)",
-                color: "var(--color-text-danger-secondary)",
-                border: "1px solid var(--color-danger-border)",
-                borderRadius: "4px",
-                cursor: "pointer",
-              }}
-            >
-              Edit Manually
-            </button>
-          </div>
-        </div>
-        );
-      })()}
-      {preview?.status === "graph-error" && (
-        <div
-          style={{
-            padding: "16px",
-            backgroundColor: "var(--color-danger-bg)",
-            borderRadius: "6px",
-            marginBottom: "16px",
-          }}
-        >
-          <div style={{ color: "var(--color-text-danger)", fontWeight: 600, marginBottom: "8px" }}>
-            ⚠️ Graph Error
-          </div>
-          <p style={{ color: "var(--color-text-danger-secondary)", fontSize: "14px", margin: "0 0 12px" }}>
-            The extracted graph DSL is invalid.
-          </p>
-          <button
-            onClick={() => setCurrentStep("editing")}
-            style={{
-              padding: "8px 16px",
-              backgroundColor: "var(--color-danger)",
-              color: "white",
-              border: "none",
-              borderRadius: "4px",
-              cursor: "pointer",
-            }}
-          >
-            Edit Manually
-          </button>
-        </div>
-      )}
-      {error && (
-        <div
-          style={{
-            padding: "12px 16px",
-            backgroundColor: "var(--color-danger-bg)",
-            borderRadius: "6px",
-            color: "var(--color-text-danger)",
-            fontSize: "14px",
-          }}
-        >
-          {error}
-        </div>
-      )}
-    </div>
-  );
-}
-
-interface EditingStepProps {
-  previewId: string;
-  preview: IngestionPreview | null;
-  formData: WizardFormData;
-  onFieldChange: (field: keyof WizardFormData, value: string | string[]) => void;
-  graphError: string;
-  onClearGraphError: () => void;
-  onGraphError: (errorMsg: string) => void;
-  tagSuggestions: string[];
-  error: string;
-  setCurrentStep: (step: WizardStep) => void;
-  onCancel: (() => void) | undefined;
-}
-
-function EditingStep({
-  previewId,
-  preview,
-  formData,
-  onFieldChange,
-  graphError,
-  onClearGraphError,
-  onGraphError,
-  tagSuggestions,
-  error,
-  setCurrentStep,
-  onCancel,
-}: EditingStepProps) {
-  return (
-    <div style={{ padding: "32px" }}>
-      <h3 style={{ margin: "0 0 24px", fontSize: "18px", fontWeight: 600 }}>
-        Edit Problem Details
-      </h3>
-
-      {previewId && preview?.sourceImage && (
-        <div style={{ marginBottom: "24px" }}>
-          <label
-            style={{
-              display: "block",
-              marginBottom: "6px",
-              fontSize: "14px",
-              fontWeight: 500,
-              color: "var(--color-text)",
-            }}
-          >
-            Original Image
-          </label>
-          <img
-            src={`/api/v1/ingestion-previews/${previewId}/image`}
-            alt="Original problem image"
-            style={{
-              maxWidth: "100%",
-              maxHeight: "300px",
-              borderRadius: "6px",
-              border: "1px solid var(--color-border)",
-            }}
-            data-testid="source-image"
-          />
-        </div>
-      )}
-
-      <div style={{ marginBottom: "24px" }}>
-        <label
-          style={{
-            display: "block",
-            marginBottom: "6px",
-            fontSize: "14px",
-            fontWeight: 500,
-            color: "var(--color-text)",
-          }}
-        >
-          Problem Text
-        </label>
-        <textarea
-          value={formData.text}
-          onChange={(e) => onFieldChange("text", e.target.value)}
-          rows={4}
-          style={{
-            width: "100%",
-            padding: "10px 12px",
-            border: "1px solid var(--color-border-muted)",
-            borderRadius: "6px",
-            fontSize: "14px",
-            fontFamily: "inherit",
-            resize: "vertical",
-            boxSizing: "border-box",
-            backgroundColor: "var(--color-surface)",
-            color: "var(--color-text)",
-          }}
-          placeholder="Enter the problem statement..."
-          data-testid="text-input"
-        />
-      </div>
-
-      <div style={{ marginBottom: "24px" }}>
-        <label
-          style={{
-            display: "block",
-            marginBottom: "6px",
-            fontSize: "14px",
-            fontWeight: 500,
-            color: "var(--color-text)",
-          }}
-        >
-          Subject
-        </label>
-        <select
-          value={formData.subject}
-          onChange={(e) => onFieldChange("subject", e.target.value)}
-          style={{
-            width: "100%",
-            padding: "10px 12px",
-            border: "1px solid var(--color-border-muted)",
-            borderRadius: "6px",
-            fontSize: "14px",
-            fontFamily: "inherit",
-            boxSizing: "border-box",
-            backgroundColor: "var(--color-surface)",
-            color: "var(--color-text)",
-          }}
-          data-testid="subject-input"
-        >
-          <option value="math">Math</option>
-          <option value="english">English</option>
-        </select>
-      </div>
-
-      <div style={{ marginBottom: "24px" }}>
-        <label
-          style={{
-            display: "block",
-            marginBottom: "6px",
-            fontSize: "14px",
-            fontWeight: 500,
-            color: "var(--color-text)",
-          }}
-        >
-          Problem Type
-        </label>
-        <select
-          value={formData.problemType}
-          onChange={(e) => onFieldChange("problemType", e.target.value)}
-          style={{
-            width: "100%",
-            padding: "10px 12px",
-            border: "1px solid var(--color-border-muted)",
-            borderRadius: "6px",
-            fontSize: "14px",
-            fontFamily: "inherit",
-            boxSizing: "border-box",
-            backgroundColor: "var(--color-surface)",
-            color: "var(--color-text)",
-          }}
-          data-testid="problem-type-input"
-        >
-          <option value="">Select a problem type…</option>
-          <option value="single-choice">Single Choice</option>
-          <option value="multi-choice">Multi Choice</option>
-          <option value="fill-in-the-blank">Fill in the Blank</option>
-          <option value="short-answer">Short Answer</option>
-        </select>
-      </div>
-
-      <div style={{ marginBottom: "24px" }}>
-        <label
-          style={{
-            display: "block",
-            marginBottom: "6px",
-            fontSize: "14px",
-            fontWeight: 500,
-            color: "var(--color-text)",
-          }}
-        >
-          Graph DSL
-        </label>
-        <textarea
-          value={formData.graphDsl}
-          onChange={(e) => onFieldChange("graphDsl", e.target.value)}
-          rows={6}
-          style={{
-            width: "100%",
-            padding: "10px 12px",
-            border: "1px solid var(--color-border-muted)",
-            borderRadius: "6px",
-            fontSize: "13px",
-            fontFamily: "monospace",
-            resize: "vertical",
-            boxSizing: "border-box",
-            backgroundColor: "var(--color-surface)",
-            color: "var(--color-text)",
-          }}
-          placeholder="Enter JSXGraph DSL code..."
-          data-testid="graph-dsl-input"
-        />
-
-        {formData.graphDsl && (
-          <div style={{ marginTop: "16px" }}>
-            <div
-              style={{
-                fontSize: "14px",
-                fontWeight: 500,
-                color: "var(--color-text)",
-                marginBottom: "8px",
-              }}
-            >
-              Graph Preview
-            </div>
-            <GraphSandbox
-              dsl={formData.graphDsl}
-              height={300}
-              onError={onGraphError}
-              onRender={onClearGraphError}
-            />
-            {graphError && (
-              <div style={{ marginTop: "8px", color: "var(--color-text-danger)", fontSize: "14px" }}>
-                {graphError}
-              </div>
-            )}
-          </div>
-        )}
-      </div>
-
-      <div style={{ marginBottom: "24px" }}>
-        <label
-          style={{
-            display: "block",
-            marginBottom: "6px",
-            fontSize: "14px",
-            fontWeight: 500,
-            color: "var(--color-text)",
-          }}
-        >
-          Correct Answer
-        </label>
-        <input
-          type="text"
-          value={formData.correctAnswer}
-          onChange={(e) => onFieldChange("correctAnswer", e.target.value)}
-          style={{
-            width: "100%",
-            padding: "10px 12px",
-            border: "1px solid var(--color-border-muted)",
-            borderRadius: "6px",
-            fontSize: "14px",
-            fontFamily: "inherit",
-            boxSizing: "border-box",
-            backgroundColor: "var(--color-surface)",
-            color: "var(--color-text)",
-          }}
-          placeholder="Enter the correct answer..."
-          data-testid="correct-answer-input"
-        />
-      </div>
-
-      {(formData.problemType === "single-choice" || formData.problemType === "multi-choice") && (
-        <div style={{ marginBottom: "24px" }}>
-          <label
-            style={{
-              display: "block",
-              marginBottom: "6px",
-              fontSize: "14px",
-              fontWeight: 500,
-              color: "var(--color-text)",
-            }}
-          >
-            Detected Choices
-          </label>
-          {(() => {
-            const choices = parseOptions(formData.text);
-            return choices.length > 0 ? (
-              <div data-testid="detected-choices" style={{ padding: "10px 12px", backgroundColor: "var(--color-surface-muted)", borderRadius: "6px", border: "1px solid var(--color-border)" }}>
-                <div data-testid="detected-choices-count" style={{ fontSize: "12px", color: "var(--color-text-muted)", marginBottom: "8px" }}>
-                  {choices.length} choice{choices.length !== 1 ? "s" : ""} detected
-                </div>
-                <div style={{ display: "flex", flexDirection: "column", gap: "4px" }}>
-                  {choices.map((choice, index) => (
-                    <div key={index} style={{ fontSize: "14px", color: "var(--color-text)", padding: "2px 0" }}>
-                      {choice}
-                    </div>
-                  ))}
-                </div>
-              </div>
-            ) : (
-              <div data-testid="detected-choices" style={{ padding: "10px 12px", backgroundColor: "var(--color-surface-muted)", borderRadius: "6px", border: "1px solid var(--color-border)", fontSize: "14px", color: "var(--color-text-muted)" }}>
-                No choices detected in problem text
-              </div>
-            );
-          })()}
-        </div>
-      )}
-
-      <div style={{ marginBottom: "24px" }}>
-        <TagInput
-          tags={formData.tags}
-          onChange={(tags) => onFieldChange("tags", tags)}
-          suggestions={tagSuggestions}
-          placeholder="Add a tag..."
-          testId="tags-input"
-        />
-      </div>
-
-      {error && (
-        <div
-          style={{
-            padding: "12px 16px",
-            backgroundColor: "var(--color-danger-bg)",
-            borderRadius: "6px",
-            color: "var(--color-text-danger)",
-            fontSize: "14px",
-            marginBottom: "16px",
-          }}
-        >
-          {error}
-        </div>
-      )}
-
-      <div style={{ display: "flex", gap: "12px" }}>
-        <button
-          onClick={() => setCurrentStep("confirming")}
-          style={{
-            padding: "10px 20px",
-            backgroundColor: "var(--color-primary)",
-            color: "white",
-            border: "none",
-            borderRadius: "6px",
-            cursor: "pointer",
-            fontSize: "14px",
-            fontWeight: 500,
-          }}
-          data-testid="review-button"
-        >
-          Review & Confirm
-        </button>
-        <button
-          onClick={onCancel}
-          style={{
-            padding: "10px 20px",
-            backgroundColor: "var(--color-surface)",
-            color: "var(--color-text)",
-            border: "1px solid var(--color-border-muted)",
-            borderRadius: "6px",
-            cursor: "pointer",
-            fontSize: "14px",
-            fontWeight: 500,
-          }}
-        >
-          Cancel
-        </button>
-      </div>
-    </div>
-  );
-}
-
-interface ConfirmingStepProps {
-  formData: WizardFormData;
-  graphError: string;
-  onGraphError: (errorMsg: string) => void;
-  error: string;
-  isLoading: boolean;
-  onConfirm: () => void;
-  setCurrentStep: (step: WizardStep) => void;
-}
-
-function ConfirmingStep({
-  formData,
-  graphError,
-  onGraphError,
-  error,
-  isLoading,
-  onConfirm,
-  setCurrentStep,
-}: ConfirmingStepProps) {
-  return (
-    <div style={{ padding: "32px" }}>
-      <h3 style={{ margin: "0 0 24px", fontSize: "18px", fontWeight: 600 }}>
-        Confirm Problem
-      </h3>
-
-      <div
-        style={{
-          backgroundColor: "var(--color-surface-muted)",
-          borderRadius: "8px",
-          padding: "24px",
-          marginBottom: "24px",
-        }}
-      >
-        <div style={{ marginBottom: "16px" }}>
-          <div
-            style={{
-              fontSize: "12px",
-              fontWeight: 600,
-              color: "var(--color-text-muted)",
-              textTransform: "uppercase",
-              letterSpacing: "0.05em",
-              marginBottom: "4px",
-            }}
-          >
-            Problem Text
-          </div>
-          <LatexText
-            text={formData.text || "(empty)"}
-            style={{ fontSize: "14px", color: "var(--color-text)" }}
-          />
-        </div>
-
-        <div style={{ marginBottom: "16px" }}>
-          <div
-            style={{
-              fontSize: "12px",
-              fontWeight: 600,
-              color: "var(--color-text-muted)",
-              textTransform: "uppercase",
-              letterSpacing: "0.05em",
-              marginBottom: "4px",
-            }}
-          >
-            Type
-          </div>
-          <div style={{ fontSize: "14px", color: "var(--color-text)" }}>
-            {formData.problemType || "(empty)"}
-          </div>
-        </div>
-
-        <div style={{ marginBottom: "16px" }}>
-          <div
-            style={{
-              fontSize: "12px",
-              fontWeight: 600,
-              color: "var(--color-text-muted)",
-              textTransform: "uppercase",
-              letterSpacing: "0.05em",
-              marginBottom: "4px",
-            }}
-          >
-            Subject
-          </div>
-          <div style={{ fontSize: "14px", color: "var(--color-text)" }}>
-            {formData.subject || "(empty)"}
-          </div>
-        </div>
-
-        <div style={{ marginBottom: "16px" }}>
-          <div
-            style={{
-              fontSize: "12px",
-              fontWeight: 600,
-              color: "var(--color-text-muted)",
-              textTransform: "uppercase",
-              letterSpacing: "0.05em",
-              marginBottom: "4px",
-            }}
-          >
-            Graph DSL
-          </div>
-          <div
-            style={{
-              fontSize: "13px",
-              color: "var(--color-text)",
-              fontFamily: "monospace",
-              backgroundColor: "var(--color-surface)",
-              padding: "8px",
-              borderRadius: "4px",
-              border: "1px solid var(--color-border)",
-            }}
-          >
-            {formData.graphDsl || "(empty)"}
-          </div>
-        </div>
-
-        <div style={{ marginBottom: "16px" }}>
-          <div
-            style={{
-              fontSize: "12px",
-              fontWeight: 600,
-              color: "var(--color-text-muted)",
-              textTransform: "uppercase",
-              letterSpacing: "0.05em",
-              marginBottom: "4px",
-            }}
-          >
-            Correct Answer
-          </div>
-          <div style={{ fontSize: "14px", color: "var(--color-text)" }}>
-            {formData.correctAnswer || "(empty)"}
-          </div>
-        </div>
-
-        {(formData.problemType === "single-choice" || formData.problemType === "multi-choice") && (
-          <div style={{ marginBottom: "16px" }}>
-            <div
-              style={{
-                fontSize: "12px",
-                fontWeight: 600,
-                color: "var(--color-text-muted)",
-                textTransform: "uppercase",
-                letterSpacing: "0.05em",
-                marginBottom: "4px",
-              }}
-            >
-              Detected Choices
-            </div>
-            {(() => {
-              const choices = parseOptions(formData.text);
-              return choices.length > 0 ? (
-                <div data-testid="detected-choices" style={{ fontSize: "14px", color: "var(--color-text)" }}>
-                  <span data-testid="detected-choices-count">{choices.length} choice{choices.length !== 1 ? "s" : ""} detected</span>
-                  <div style={{ marginTop: "4px", display: "flex", flexDirection: "column", gap: "2px" }}>
-                    {choices.map((choice, index) => (
-                      <div key={index}>{choice}</div>
-                    ))}
-                  </div>
-                </div>
-              ) : (
-                <div data-testid="detected-choices" style={{ fontSize: "14px", color: "var(--color-text-muted)" }}>
-                  No choices detected
-                </div>
-              );
-            })()}
-          </div>
-        )}
-
-        <div>
-          <div
-            style={{
-              fontSize: "12px",
-              fontWeight: 600,
-              color: "var(--color-text-muted)",
-              textTransform: "uppercase",
-              letterSpacing: "0.05em",
-              marginBottom: "4px",
-            }}
-          >
-            Tags
-          </div>
-          <div style={{ fontSize: "14px", color: "var(--color-text)" }}>
-            {formData.tags.length > 0 ? formData.tags.join(", ") : "(empty)"}
-          </div>
-        </div>
-      </div>
-
-      {formData.graphDsl && (
-        <div style={{ marginBottom: "24px" }}>
-          <div
-            style={{
-              fontSize: "14px",
-              fontWeight: 500,
-              color: "var(--color-text)",
-              marginBottom: "8px",
-            }}
-          >
-            Graph Preview
-          </div>
-          <GraphSandbox
-            dsl={formData.graphDsl}
-            height={300}
-            onError={onGraphError}
-          />
-          {graphError && (
-            <div style={{ marginTop: "8px", color: "var(--color-text-danger)", fontSize: "14px" }}>
-              {graphError}
-            </div>
-          )}
-        </div>
-      )}
-
-      {error && (
-        <div
-          style={{
-            padding: "12px 16px",
-            backgroundColor: "var(--color-danger-bg)",
-            borderRadius: "6px",
-            color: "var(--color-text-danger)",
-            fontSize: "14px",
-            marginBottom: "16px",
-          }}
-        >
-          {error}
-        </div>
-      )}
-
-      <div style={{ display: "flex", gap: "12px" }}>
-        <button
-          onClick={onConfirm}
-          disabled={isLoading}
-          style={{
-            padding: "10px 20px",
-            backgroundColor: "var(--color-success)",
-            color: "white",
-            border: "none",
-            borderRadius: "6px",
-            cursor: isLoading ? "not-allowed" : "pointer",
-            fontSize: "14px",
-            fontWeight: 500,
-            opacity: isLoading ? 0.7 : 1,
-          }}
-          data-testid="confirm-button"
-        >
-          {isLoading ? "Creating..." : "Confirm & Save"}
-        </button>
-        <button
-          onClick={() => setCurrentStep("editing")}
-          disabled={isLoading}
-          style={{
-            padding: "10px 20px",
-            backgroundColor: "var(--color-surface)",
-            color: "var(--color-text)",
-            border: "1px solid var(--color-border-muted)",
-            borderRadius: "6px",
-            cursor: isLoading ? "not-allowed" : "pointer",
-            fontSize: "14px",
-            fontWeight: 500,
-          }}
-        >
-          Back to Edit
-        </button>
-      </div>
-    </div>
   );
 }
